app/requirements.txt
amqp==5.3.1
asgiref==3.8.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
chardet==3.0.4
charset-normalizer==2.0.7
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.9
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
dotenv==0.9.9
drf-yasg==1.21.9
filelock==3.17.0
inflection==0.5.1
kombu==5.4.2
packaging==24.2
paramiko==3.5.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
ssh-pymongo==1.0.4
sshtunnel==0.4.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.25.11
vine==5.1.0
virtualenv==20.29.1
wcwidth==0.2.13